+++ src/hooks/auth/index.ts
@@ -2,4 +2,6 @@ export * from "./use-google-logged-in";
 export * from "./use-cookies";
 export * from "./use-logout";
 export * from "./use-profile";
-export * from "./use-signup";
+export * from "./use-email";
+export * from "./use-password";
+export * from "./use-verification-code";
