+++ src/hooks/auth/use-email.ts
@@ -1,0 +1,20 @@
+import { useState } from "react";
+
+export function useEmail() {
+  const [email, setEmail] = useState("");
+  const [isEmailValid, setIsEmailValid] = useState(false);
+
+  const handleEmailChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const value = e.target.value;
+    setEmail(value);
+
+    const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+    setIsEmailValid(emailRegex.test(value));
+  };
+
+  return {
+    email,
+    isEmailValid,
+    handleEmailChange,
+  };
+}
