+++ src/hooks/auth/use-password.ts
@@ -1,0 +1,26 @@
+import { useState } from "react";
+
+export function usePassword() {
+  const [password, setPassword] = useState("");
+  const [confirmPassword, setConfirmPassword] = useState("");
+  const [passwordError, setPasswordError] = useState("");
+
+  const handleConfirmPasswordChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const value = e.target.value;
+    setConfirmPassword(value);
+
+    if (password && value !== password) {
+      setPasswordError("비밀번호가 일치하지 않습니다.");
+    } else {
+      setPasswordError("");
+    }
+  };
+
+  return {
+    password,
+    confirmPassword,
+    passwordError,
+    setPassword,
+    handleConfirmPasswordChange,
+  };
+}
