+++ src/hooks/auth/use-verification-code.ts
@@ -1,0 +1,19 @@
+import { useState } from "react";
+
+export function useVerificationCode() {
+  const [verificationCode, setVerificationCode] = useState("");
+  const [isCodeVerified, setIsCodeVerified] = useState(false);
+
+  const handleVerification = () => {
+    if (verificationCode.trim() !== "") {
+      setIsCodeVerified(true);
+    }
+  };
+
+  return {
+    verificationCode,
+    isCodeVerified,
+    setVerificationCode,
+    handleVerification,
+  };
+}
