+++ src/pages/signup-page.tsx
@@ -3,23 +3,18 @@ import PasswordSection from "@/components/signup/signup-password-section";
 import SubmitButton from "@/components/signup/signup-submit-button";
 import SignupFooter from "@/components/signup/signup-footer";
 import EmailSection from "@/components/signup/signup-email-section";
-import useSignup from "@/hooks/auth/use-signup";
+
+import { useEmail } from "@/hooks/auth/use-email";
+import { usePassword } from "@/hooks/auth/use-password";
+import { useVerificationCode } from "@/hooks/auth/use-verification-code";
 
 export default function SignupPage() {
-  const {
-    email,
-    isEmailValid,
-    verificationCode,
-    isCodeVerified,
-    password,
-    confirmPassword,
-    passwordError,
-    setVerificationCode,
-    setPassword,
-    handleEmailChange,
-    handleVerification,
-    handleConfirmPasswordChange,
-  } = useSignup();
+  // 각각 역할별로 분리해서 훅 가져오기
+  const { email, isEmailValid, handleEmailChange } = useEmail();
+  const { verificationCode, isCodeVerified, setVerificationCode, handleVerification } =
+    useVerificationCode();
+  const { password, confirmPassword, passwordError, setPassword, handleConfirmPasswordChange } =
+    usePassword();
 
   return (
     <>
@@ -34,6 +29,7 @@ export default function SignupPage() {
           handleVerification={handleVerification}
           setVerificationCode={setVerificationCode}
         />
+
         {isCodeVerified && (
           <>
             <PasswordSection
